+++ src/check.rs
@@ -339,7 +339,7 @@ impl<'a> VirtualEnv<'a> {
                     let sig = self.pop_func()?.signature();
                     self.handle_args_outputs(2, sig.outputs)?;
                 }
-                Spawn => {
+                Spawn | PoolSpawn => {
                     let sig = self.pop_func()?.signature();
                     self.handle_args_outputs(sig.args, 1)?;
                 }